Move the SwapDigest definition to the swap_digest module
