+++ comit/src/network/swap_digest.rs
@@ -1,5 +1,8 @@
-use crate::{asset, identity, network::SwapDigest, RelativeTime, Timestamp};
-use digest::Digest;
+use crate::{asset, identity, RelativeTime, Timestamp};
+use digest::{Digest, ToDigestInput};
+use libp2p::multihash::{self, Multihash};
+use serde::{de::Error, Deserialize, Deserializer, Serialize, Serializer};
+use std::fmt;
 
 pub fn herc20_halbit<S: Into<Herc20Halbit>>(swap: S) -> SwapDigest {
     swap.into().digest()
@@ -83,4 +86,55 @@ pub struct HbitHerc20 {
     pub erc20_amount: asset::Erc20Quantity,
     #[digest(prefix = "3003")]
     pub token_contract: identity::Ethereum,
+}
+
+#[derive(Clone, Debug, PartialOrd, Ord, PartialEq, Eq, Hash)]
+pub struct SwapDigest(Multihash);
+
+impl SwapDigest {
+    pub fn new(multihash: Multihash) -> Self {
+        Self(multihash)
+    }
+}
+
+impl ToDigestInput for SwapDigest {
+    fn to_digest_input(&self) -> Vec<u8> {
+        self.0.clone().into_bytes()
+    }
+}
+
+impl digest::Hash for SwapDigest {
+    fn hash(bytes: &[u8]) -> Self {
+        Self(multihash::Sha3_256::digest(bytes))
+    }
+}
+
+impl fmt::Display for SwapDigest {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", hex::encode(self.0.as_bytes()))
+    }
+}
+
+impl Serialize for SwapDigest {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let hex = hex::encode(self.0.as_bytes());
+
+        serializer.serialize_str(&hex)
+    }
+}
+
+impl<'de> Deserialize<'de> for SwapDigest {
+    fn deserialize<D>(deserializer: D) -> Result<Self, <D as Deserializer<'de>>::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let hex = String::deserialize(deserializer)?;
+        let bytes = hex::decode(hex).map_err(D::Error::custom)?;
+        let multihash = multihash::Multihash::from_bytes(bytes).map_err(D::Error::custom)?;
+
+        Ok(SwapDigest::new(multihash))
+    }
 }
